Refactor playlist detail model and view
This commit refactors the `PlaylistDetailModel` class in `playlist_detail_model.py` to use the `content_info_items` parameter instead of `content`. It also updates the `playlist_detail` view in `views.py` to sort the `ordered_content_items` based on `display_order` and populate the `content_items` list accordingly. This ensures that the content items are displayed in the correct order in the `playlists_detail.html` template.